CASMCMS-8029: Explicitly set permission bits during AEE setup
User set umask values apparently affect the overall run pod execution of containers deployed on worker nodes. When a user changes the default umask for the system, newly copied files inherit the permission mask as set. This prevents the ansible-execution-environment inventory layer from correctly providing a suitable key file to use to connect with nodes. The issue is further complicated by the migration to non-root containers, which provides the key file mount in variable ways based on the overall version of the CFS release. By setting the bits explicitly to known accepted values, we have a solution that works in all known releases.